Rename to read handler.
